requirements.txt
tweepy
ccxt == 1.45.18
python-dotenv
